cfa_vpn/requirements.txt
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
environs==10.3.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
gspread==6.0.0
idna==3.6
marshmallow==3.20.2
oauthlib==3.2.2
outline-vpn-api==4.0.1
packaging==23.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dotenv==1.0.1
python-http-client==3.3.7
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sendgrid==6.11.0
starkbank-ecdsa==2.2.0
StrEnum==0.4.15
urllib3==2.1.0
